bgpd: set and process route-map locator upon L3VPN exportation
Upon using a route-map which sets a locator, create a bgp srv6 locator
context on the matching bgp path. Handle the configuration changes on
the bgp path. Note that the described action works ok only when using
the route-map at vpn exportation in the vrf, like the below
configuration suggests:

> route-map rmap permit 10
>  set segment-routing srv6 locator loc1
> exit
> router bgp 65500 vrf vrf1
>  address-family ipv4 unicast
>   route-map vpn export rmap
>

Signed-off-by: Dmytro Shytyi <dmytro.shytyi@6wind.com>
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>