bgpd: add bgp srv6 locator route-map structure framework
Add a structure in each BGP instance, indexed by AFI, that stores the
list of known locators.
Add a pointer to a bgp srv6 locator structure in the bgp_path_info using
it. That structure knows the list of paths using it.

A 'show bgp locator-route-map' command to dump the references.

Signed-off-by: Dmytro Shytyi <dmytro.shytyi@6wind.com>
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>