lib, zebra: move srv6_sid_ctx structure to lib, remove useless functions
When using a route-map to allocate SIDs, only SID automatic mechanism is
granted. In the meantime, this should still be possible to do
allocation per l3vrf in an explicit way.
This commit shares the SID allocation mode on all the daemons.
The next commit will introduce allocation mode in the srv6_sid_ctx
structure.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>