*: add alloc mode to srv6 sid context
SRv6 contexts are not enough, in the context of being able to select a
specific locator, in a certain manner,  different than the default one
configured.

In the scope of using a route-map to select a specific locator in BGP,
two issues arise:
- the locator used is different
- the allocation mode may differ for a same locator.

This commit attempts to address the latter case by introducing the
allocation mode in srv6 contexts. Consequently, each daemon will
announce in the srv6 sid context whether the SID is static or
dynamic.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>